fixed aspects of myseek

+++ myio.h
@@ -3,4 +3,4 @@ int myclose(struct File *filePtr);
 int myread(struct File *filePtr, char *buf, size_t count);
 int mywrite(struct File *filePtr, char *buf, size_t count); 
 void myflush(struct File *filePtr);
-int myseek(struct File *filePtr, off_t offset, int whence); 
+int myseek(struct File *filePtr, int offset, int whence);
